refactor: move to utils file
