+++ packages/web/src/javascripts/Utils/DateUtils.ts
@@ -3,3 +3,11 @@ import { dateToLocalizedString } from '@standardnotes/snjs/'
 export const formatLastSyncDate = (lastUpdatedDate: Date) => {
   return dateToLocalizedString(lastUpdatedDate)
 }
+
+export const formatDateForContextMenu = (date: Date | undefined) => {
+  if (!date) {
+    return
+  }
+
+  return `${date.toDateString()} ${date.toLocaleTimeString()}`
+}
